Parse max level verbosity output
